requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asgiref==3.9.1
asyncpg==0.30.0
bcrypt==4.3.0
blinker==1.9.0
boto3==1.37.21
botocore==1.37.21
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dbus-python==1.4.0
Deprecated==1.2.18
distro==1.9.0
distro-info==1.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.74.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
keyring==25.6.0
launchpadlib==2.1.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.7
Mako==1.3.9
MarkupSafe==3.0.2
more-itertools==10.6.0
moto==5.1.1
oauthlib==3.2.2
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-instrumentation-asgi==0.56b0
opentelemetry-instrumentation-fastapi==0.56b0
opentelemetry-instrumentation-logging==0.56b0
opentelemetry-instrumentation-requests==0.56b0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opentelemetry-util-http==0.56b0
packaging==24.2
passlib==1.7.4
pika==1.3.2
pillow==11.1.0
pluggy==1.5.0
protobuf==6.31.1
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2
PyJWT==2.10.1
pyotp==2.9.0
pyparsing==3.2.3
pytest==8.3.5
pytest-sqlalchemy-mock==0.1.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-oauth2==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
responses==0.25.7
rsa==4.9
s3transfer==0.11.4
SecretStorage==3.3.3
setuptools==78.1.0
simple-term-menu==1.6.6
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
types-PyYAML==6.0.12.20250326
typing-inspection==0.4.1
typing_extensions==4.13.0
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
wadllib==2.0.0
Werkzeug==3.1.3
wrapt==1.17.2
xmltodict==0.14.2
zipp==3.21.0
